requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.38.0
sentence-transformers>=2.6.0
faiss-cpu>=1.7.4          

Pillow>=10.2.0
requests>=2.31.0

pandas>=2.1.0
numpy>=1.26.0

fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9   
pydantic>=2.6.0

streamlit>=1.32.0

python-dotenv>=1.0.0
tqdm>=4.66.0
accelerate>=0.27.0
